fix: replace `unsigned int` with `size_t` everywhere.

+++ libft/bzero.c
@@ -1,7 +1,9 @@
-void	ft_bzero(void *s, unsigned int n)
+#include "libft.h"
+
+void	ft_bzero(void *s, size_t n)
 {
 	unsigned char	*string;
-	unsigned int	i;
+	size_t	i;
 
 	string = s;
 	i = 0;
